+++ src/Commands/ProcessMetrics.php
@@ -4,6 +4,7 @@ namespace OneMediaLabs\MixpostMcp\Commands;
 
 use Illuminate\Console\Command;
 use OneMediaLabs\MixpostMcp\Concerns\AccountsOption;
+use OneMediaLabs\MixpostMcp\SocialProviders\LinkedIn\Jobs\ProcessLinkedInMetricsJob;
 use OneMediaLabs\MixpostMcp\SocialProviders\Mastodon\Jobs\ProcessMastodonMetricsJob;
 use OneMediaLabs\MixpostMcp\SocialProviders\Twitter\Jobs\ProcessTwitterMetricsJob;
 
@@ -21,6 +22,7 @@ class ProcessMetrics extends Command
             $job = match ($account->provider) {
                 'twitter' => ProcessTwitterMetricsJob::class,
                 'mastodon' => ProcessMastodonMetricsJob::class,
+                'linkedin' => ProcessLinkedInMetricsJob::class,
                 default => null,
             };